# Conflicts:
#	WMS/Data/Core/Models/BaseModel.cs
#	WMS/Data/WebAPI.Contracts/ServiceContracts.autogen.cs

+++ Common/DataModels/CellHeightClass.cs
@@ -13,9 +13,9 @@ namespace Ferretto.Common.DataModels
 
         public int Id { get; set; }
 
-        public int MaxHeight { get; set; }
+        public double MaxHeight { get; set; }
 
-        public int MinHeight { get; set; }
+        public double MinHeight { get; set; }
 
         #endregion
     }
