+++ WMS/Data/WebAPI.Contracts/ServiceContracts.autogen.cs
@@ -346,11 +346,11 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         System.Threading.Tasks.Task<System.Collections.ObjectModel.ObservableCollection<AllowedItemInCompartment>> GetAllowedItemsAsync(int id, System.Threading.CancellationToken cancellationToken);
     
         /// <exception cref="SwaggerException">A server side error occurred.</exception>
-        System.Threading.Tasks.Task<int?> GetMaxCapacityAsync(int width, int height, int itemId);
+        System.Threading.Tasks.Task<int?> GetMaxCapacityAsync(double width, double height, int itemId);
     
         /// <exception cref="SwaggerException">A server side error occurred.</exception>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        System.Threading.Tasks.Task<int?> GetMaxCapacityAsync(int width, int height, int itemId, System.Threading.CancellationToken cancellationToken);
+        System.Threading.Tasks.Task<int?> GetMaxCapacityAsync(double width, double height, int itemId, System.Threading.CancellationToken cancellationToken);
     
         /// <exception cref="SwaggerException">A server side error occurred.</exception>
         System.Threading.Tasks.Task<System.Collections.ObjectModel.ObservableCollection<object>> GetUniqueValuesAsync(string propertyName);
@@ -1230,13 +1230,13 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public string CellTypeDescription { get; set; }
     
         [Newtonsoft.Json.JsonProperty("XCoordinate", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? XCoordinate { get; set; }
+        public double? XCoordinate { get; set; }
     
         [Newtonsoft.Json.JsonProperty("YCoordinate", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? YCoordinate { get; set; }
+        public double? YCoordinate { get; set; }
     
         [Newtonsoft.Json.JsonProperty("ZCoordinate", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? ZCoordinate { get; set; }
+        public double? ZCoordinate { get; set; }
     
         public string ToJson() 
         {
@@ -1387,13 +1387,13 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public Side Side { get; set; }
     
         [Newtonsoft.Json.JsonProperty("XCoordinate", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? XCoordinate { get; set; }
+        public double? XCoordinate { get; set; }
     
         [Newtonsoft.Json.JsonProperty("YCoordinate", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? YCoordinate { get; set; }
+        public double? YCoordinate { get; set; }
     
         [Newtonsoft.Json.JsonProperty("ZCoordinate", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? ZCoordinate { get; set; }
+        public double? ZCoordinate { get; set; }
     
         public string ToJson() 
         {
@@ -1445,7 +1445,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int? HandlingParametersCorrection { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Always)]
-        public int Height { get; set; }
+        public double Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("InCycleCount", Required = Newtonsoft.Json.Required.Always)]
         public int InCycleCount { get; set; }
@@ -1469,7 +1469,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public System.DateTime? LastStoreDate { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Length", Required = Newtonsoft.Json.Required.Always)]
-        public int Length { get; set; }
+        public double Length { get; set; }
     
         [Newtonsoft.Json.JsonProperty("LoadingUnitStatusDescription", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string LoadingUnitStatusDescription { get; set; }
@@ -1504,7 +1504,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int Weight { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Width", Required = Newtonsoft.Json.Required.Always)]
-        public int Width { get; set; }
+        public double Width { get; set; }
     
         public string ToJson() 
         {
@@ -1615,7 +1615,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public bool HasRotation { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Height { get; set; }
+        public double? Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("InventoryDate", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public System.DateTime? InventoryDate { get; set; }
@@ -1681,13 +1681,13 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public string Sub2 { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Width", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Width { get; set; }
+        public double? Width { get; set; }
     
         [Newtonsoft.Json.JsonProperty("XPosition", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? XPosition { get; set; }
+        public double? XPosition { get; set; }
     
         [Newtonsoft.Json.JsonProperty("YPosition", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? YPosition { get; set; }
+        public double? YPosition { get; set; }
     
         public string ToJson() 
         {
@@ -1711,7 +1711,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public bool HasRotation { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Height { get; set; }
+        public double? Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("IsItemPairingFixed", Required = Newtonsoft.Json.Required.Always)]
         public bool IsItemPairingFixed { get; set; }
@@ -1747,13 +1747,13 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public string Sub2 { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Width", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Width { get; set; }
+        public double? Width { get; set; }
     
         [Newtonsoft.Json.JsonProperty("XPosition", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? XPosition { get; set; }
+        public double? XPosition { get; set; }
     
         [Newtonsoft.Json.JsonProperty("YPosition", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? YPosition { get; set; }
+        public double? YPosition { get; set; }
     
         public string ToJson() 
         {
@@ -1828,10 +1828,10 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
     public partial class CompartmentType : BaseModelOfInt32
     {
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Height { get; set; }
+        public double? Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Width", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Width { get; set; }
+        public double? Width { get; set; }
     
         public string ToJson() 
         {
@@ -2297,7 +2297,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int? FifoTimeStore { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Height { get; set; }
+        public double? Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Image", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string Image { get; set; }
@@ -2324,7 +2324,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public System.DateTime? LastStoreDate { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Length", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Length { get; set; }
+        public double? Length { get; set; }
     
         [Newtonsoft.Json.JsonProperty("ManagementType", Required = Newtonsoft.Json.Required.Always)]
         [System.ComponentModel.DataAnnotations.Required(AllowEmptyStrings = true)]
@@ -2356,7 +2356,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int TotalAvailable { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Width", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Width { get; set; }
+        public double? Width { get; set; }
     
         public string ToJson() 
         {
@@ -2413,7 +2413,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int? FifoTimeStore { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Height { get; set; }
+        public double? Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Image", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string Image { get; set; }
@@ -2440,7 +2440,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public System.DateTime? LastStoreDate { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Length", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Length { get; set; }
+        public double? Length { get; set; }
     
         [Newtonsoft.Json.JsonProperty("ManagementType", Required = Newtonsoft.Json.Required.Always)]
         [System.ComponentModel.DataAnnotations.Required(AllowEmptyStrings = true)]
@@ -2481,7 +2481,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int TotalStock { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Width", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? Width { get; set; }
+        public double? Width { get; set; }
     
         public string ToJson() 
         {
@@ -2663,7 +2663,7 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int? HandlingParametersCorrection { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Always)]
-        public int Height { get; set; }
+        public double Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("InCycleCount", Required = Newtonsoft.Json.Required.Always)]
         public int InCycleCount { get; set; }
@@ -2759,16 +2759,16 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
     public partial class LoadingUnitSize 
     {
         [Newtonsoft.Json.JsonProperty("Height", Required = Newtonsoft.Json.Required.Always)]
-        public int Height { get; set; }
+        public double Height { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Length", Required = Newtonsoft.Json.Required.Always)]
-        public int Length { get; set; }
+        public double Length { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Weight", Required = Newtonsoft.Json.Required.Always)]
         public int Weight { get; set; }
     
         [Newtonsoft.Json.JsonProperty("Width", Required = Newtonsoft.Json.Required.Always)]
-        public int Width { get; set; }
+        public double Width { get; set; }
     
         public string ToJson() 
         {
@@ -3055,10 +3055,10 @@ namespace Ferretto.WMS.Data.WebAPI.Contracts
         public int? CompartmentId { get; set; }
     
         [Newtonsoft.Json.JsonProperty("CompartmentTypeHeight", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? CompartmentTypeHeight { get; set; }
+        public double? CompartmentTypeHeight { get; set; }
     
         [Newtonsoft.Json.JsonProperty("CompartmentTypeWidth", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? CompartmentTypeWidth { get; set; }
+        public double? CompartmentTypeWidth { get; set; }
     
         [Newtonsoft.Json.JsonProperty("CreationDate", Required = Newtonsoft.Json.Required.Always)]
         [System.ComponentModel.DataAnnotations.Required(AllowEmptyStrings = true)]
